Added the search box
Changed index.html to index.php, so you’ll need to view it on a server
instead of just locally.
I made the search icon a search submit button, and thus moved it to the
right because that’s usually where a submit button goes. However, it
can be easily moved to the left.